+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-ejb/src/java/org/dcm4chex/archive/ejb/interfaces/FileDTO.java
@@ -29,7 +29,7 @@ public final class FileDTO implements Serializable {
 
     private String tsuid;
 
-    private int size;
+    private long size;
 
     private byte[] md5;
 
@@ -106,14 +106,14 @@ public final class FileDTO implements Serializable {
     /**
      * @return Returns the size.
      */
-    public final int getFileSize() {
+    public final long getFileSize() {
         return size;
     }
 
     /**
      * @param size The size to set.
      */
-    public final void setFileSize(int size) {
+    public final void setFileSize(long size) {
         this.size = size;
     }
